updated groups BE

+++ groupmeet-backend/src/controllers/groupController.js
@@ -26,9 +26,13 @@ const getGroup = async (req, res) => {
 // create new group
 const createGroup = async (req, res) => {
     const {name, members} = req.body
+    const creator = members[0];
     
     try {
-      const group = await Group.create({name, members})
+      const group = await Group.create({name, members,creator})
+      const user = await User.findOneAndUpdate({email: creator}, {
+        $push:{groups: group}
+      })
       res.status(200).json(group)
     } catch (error) {
       res.status(400).json({error: error.message})
@@ -37,8 +41,20 @@ const createGroup = async (req, res) => {
 
 // delete a group
 const deleteGroup = async (req, res) => {
-    const { id } = req.params
-  
+    const { id, cur_usr } = req.params
+
+    const target_group = await Group.find({name: id})
+
+    if(target_group[0]['creator'] != cur_usr) {
+      return res.status(404).json({error: 'You need admin rights to delete this group'})
+    }
+
+    let members = target_group[0]['members'];
+    for(let i=0;i<members.length;i++)
+    {
+      const user = await User.findOneAndUpdate({ email:members[i] }, 
+        { $pullAll: { groups: [target_group[0]._id] } } )
+    }
     const group = await Group.findOneAndDelete({name: id})
   
     if(!group || group.length == 0) {
@@ -52,6 +68,12 @@ const deleteGroup = async (req, res) => {
 const deleteGroupMember = async (req, res) => {
   const group = await Group.updateOne({ name: req.body.name }, 
     { $pullAll: { members: [req.body.members] } } )
+  
+  const target_group = await Group.find({name:req.body.name})
+  console.log(target_group)
+  const user = await User.findOneAndUpdate({ email: req.body.members }, 
+    { $pullAll: { groups: [target_group[0]._id] } } )
+
 
   if(!group || group.length == 0) {
     return res.status(400).json({error: 'ERROR'})
@@ -64,7 +86,7 @@ const deleteGroupMember = async (req, res) => {
   const updateGroup = async (req, res) => {
     const { id } = req.params
     const user = await User.find({email:req.body.members})
-
+    
     if (!user || user.length == 0) {
       return res.status(404).json({error: 'No such user found'})
     }
@@ -72,7 +94,12 @@ const deleteGroupMember = async (req, res) => {
     const group = await Group.findOneAndUpdate({name: id}, {
       $push:{members: req.body.members}
     })
-  
+    
+    const target_group = await Group.find({name: id})
+    const target_user = await User.findOneAndUpdate({email: req.body.members}, {
+      $push:{groups: target_group}
+    })
+
     if (!group || group.length == 0) {
       return res.status(400).json({error: 'No such group'})
     }
